feat: add glob pattern for finding test case files
