+++ src/app/test/utils/getLoaderVersion.ts
@@ -1,0 +1,3 @@
+export function getLoaderVersion() {
+  return `3.${Math.floor(Math.random() * 10)}.${Math.floor(Math.random() * 10)}`
+}
